Move shortcuts implementation into feature module
