+++ client/build.gradle
@@ -63,6 +63,7 @@ dependencies {
   api project(':feature-users')
   api project(':feature-settings')
   api project(':feature-about')
+  api project(':feature-shortcuts')
 
   implementation 'androidx.appcompat:appcompat:1.1.0-rc01'
   implementation 'androidx.legacy:legacy-support-v4:1.0.0'
