+++ feature-shortcuts/src/main/AndroidManifest.xml
@@ -1,0 +1,20 @@
+<manifest
+  xmlns:android="http://schemas.android.com/apk/res/android"
+  package="com.jraska.github.client.shortcuts">
+
+  <application>
+    <activity android:name=".ShortcutHandlerActivity">
+      <intent-filter>
+        <action android:name="android.intent.action.VIEW"/>
+
+        <category android:name="android.intent.category.DEFAULT"/>
+        <category android:name="android.intent.category.BROWSABLE"/>
+
+        <data
+          android:host="github.com"
+          android:pathPattern="/.*"
+          android:scheme="clientgithubjraskacom"/>
+      </intent-filter>
+    </activity>
+  </application>
+</manifest>
