+++ feature-shortcuts/src/main/java/com/jraska/github/client/shortcuts/ShortcutHandlerActivity.kt
@@ -1,11 +1,11 @@
-package com.jraska.github.client.ui
+package com.jraska.github.client.shortcuts
 
 import android.os.Bundle
 import com.jraska.github.client.core.android.BaseActivity
 import com.jraska.github.client.core.android.inputUrl
 import com.jraska.github.client.core.android.viewModel
 
-class ShortcutHandlerActivity : BaseActivity() {
+internal class ShortcutHandlerActivity : BaseActivity() {
 
   override fun onCreate(savedInstanceState: Bundle?) {
     super.onCreate(savedInstanceState)
