+++ feature-shortcuts/src/main/java/com/jraska/github/client/shortcuts/ShortcutHandlerModel.kt
@@ -1,4 +1,4 @@
-package com.jraska.github.client.ui
+package com.jraska.github.client.shortcuts
 
 import androidx.lifecycle.ViewModel
 import com.jraska.github.client.DeepLinkHandler
@@ -9,7 +9,7 @@ import com.jraska.github.client.analytics.EventAnalytics
 import okhttp3.HttpUrl
 import javax.inject.Inject
 
-class ShortcutHandlerModel @Inject constructor(
+internal class ShortcutHandlerModel @Inject constructor(
   private val deepLinkHandler: DeepLinkHandler,
   private val eventAnalytics: EventAnalytics
 ) : ViewModel() {
